+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2](https://git.cross-systems.ch/wide-front/modulus/tags/v1.0.2) - 2019/07/09
+- Remove Controller builder method
+- Add WebComponent builder method
+- Add Gitlab CI for the documentation
 
 ## [1.0.1](https://git.cross-systems.ch/wide-front/modulus/tags/v1.0.1) - 2019/07/08
 - Trap focus in open modal